+++ changelog.md
@@ -1,13 +1,14 @@
 # Vertx Utils changelog
 ## [2.1.0] - UNRELEASED
 ### Breaking Changes
-* None.
+* Updated to use Vert.x v5, which removes deprecated functions from v4, and moves to a `Future` based approach instead of callbacks.
 
 ### New Features
 * None.
 
 ### Enhancements
-* None.
+* `CATCH_ALL_API_FAILURE_HANDLER` and `CATCH_ALL_API_FAILURE_HANDLER_WITH_EXCEPTION_LOGGING` now return the status code of a failed `context`, if it is set,
+  instead of always using `500 - INTERNAL_SERVER_ERROR`.
 
 ### Fixes
 * None.