+++ changelog.md
@@ -7,14 +7,15 @@
 * None.
 
 ### Enhancements
-* `CATCH_ALL_API_FAILURE_HANDLER` and `CATCH_ALL_API_FAILURE_HANDLER_WITH_EXCEPTION_LOGGING` now return the status code of a failed `context`, if it is set,
-  instead of always using `500 - INTERNAL_SERVER_ERROR`.
+* `CATCH_ALL_API_FAILURE_HANDLER` and `CATCH_ALL_API_FAILURE_HANDLER_WITH_EXCEPTION_LOGGING` now:
+  * Return the status code of a failed `context`, if it is set, instead of always using `500 - INTERNAL_SERVER_ERROR`.
+  * Capture and ignore `ClosedChannelException` exceptions in addition to `VertxException` versions.
 
 ### Fixes
 * None.
 
 ### Notes
-* None.
+* Removed the `com.jayway.awaitility:awaitility` dependency. Its scope was `provided`, so this should have no impact.
 
 ## [2.0.0] - 2026-05-21
 ### Breaking Changes